# Optimize raw html padding for small depths